+++ internal/downloader/writer.go
@@ -11,10 +11,10 @@ func (w *DownloadWriter) Write(p []byte) (n int, err error) {
 
 func (w *DownloadWriter) UpdateProgress() {
 	for i := range w.listeners {
-		w.listeners[i](w.total)
+		w.listeners[i](uint64(w.total))
 	}
 }
 
-func (w *DownloadWriter) AddListener(listener func(progress int)) {
+func (w *DownloadWriter) AddListener(listener func(progress uint64)) {
 	w.listeners = append(w.listeners, listener)
 }
